Initial import.
The dialog library attempts to provide a fairly simplistic set of fixed-
presentation menus, input boxes, gauges, file requestors and other gen-
eral purpose GUI (a bit of a stretch, since it uses ncurses) objects.

This was supplied in pkg/18429 by Miduel Mendez <flynn@energygq.homeip.net>.
The PR can be closed as soon as this is moved to pkgsrc.