Updated to 0.7.1.
We have released GTetrinet 0.7.1, a maintenance release for 0.7.0. The biggest
thing is the new configuration option to disable the channel list feature,
which should help people having problems on servers with a non-supported /list
format. Support for the Java Tetrinet server and TetriDome has been added too,
as well as other minor bugfixes. Thanks to the GNOME translators, there are new
or updated translations, too. Have fun!