Initial import of tinylogin
TinyLogin is a suite of tiny Unix utilities for handling logging
into, being authenticated by, changing one's password for, and
otherwise maintaining users and groups on an embedded system.

It also provides shadow password support to enhace system security.
TinyLogin is, as the name implies, very small, and makes an
excellent complement to BusyBox on an embedded system (though it
can of course be used without).