Updated to 0.12
Changes:

The 0.12 release has been improved by adding support for cancelling
handlers waiting for reply. We noticed we needed that in Gossip when
for example closing a dialog waiting for a vCard reply. Loudmouth now
also checks for supported authentication schemes when trying to
authenticate (currently only supports digest and plain).