Import of a package for natali v1.25
This library is a partial implementation of the AT&T/Apple AppleTalk
Library Interface.  The ALI was designed as the standard AppleTalk API
for Unix System V Release 4.

This library is designed for use with Netatalk, a Unix AppleTalk
implementation from the University of Michigan; see
http://ppr.trincoll.edu/

The package uses a buildlink file for Netatalk which I submitted
separately.