Downloader for X is an userfriendly program with nice X interface to download files from the Internet. It suppotrs both FTP and HTTP protocols, supports resuming (of course if server side supports it) and makes downloading files from internet easy. Downloader for X is an analogue for of such applications for Windows as ReGet, Go!Zilla and GetRight.
Of course this program is not a copy of any of programs above. Downloader
for X was written from scratch in the hope to decrease a pain of downloading
many files from Internet even on really bad modem lines. But if you are
happy owner of T1 connection then D4X is for you too because It can use this
bandwidth completely! And this is not a end of the list of Downloader for
X's features!