Chora is our tool for viewing code repositories that are managed using the CVS source control system. It aims to provide a high degree of integration with CVS and the other Horde web-based tools.
Right now, Chora comes close to matching existing CVS-Web scripts feature
for feature, and provides a solid base for future development. Some exciting
new features include a visual branch viewing tool, which intelligently
displays the branch history of a given file. Branches have traditionally
been a source of confusion when using CVS, and this feature aims to make it
a bit clearer and easier to use.

From TODO:
$Id$

Prepare install target (see docs/INSTALL). As I have created a basic package
for it and then noticed it's PHP-based (and I am totally not interested in
PHP development right now), I decided to leave it, but I import it anyway,
so someone can get it working in the future.