0.8.0-402 Changes:
- The responsiveness to requests is reduced to the fingerprint's limit for all keys
but XML text keys. This now really prevents "remote scanning" of a node. Requests
which are below a node fingerprint's threshold for a certain route are not answered
or forwarded.

- The news source has undergone a not only cosmetic rewrite. There are no more calls
to scalloc() for each and every bit, but rather one scalloc() per board or message,
including all header fields. The number of entries can now grow dynamically, as long
as there's enough shared memory available

- The news layout has been changed. There's a new column "yesterday", which I won't
describe ;-) And you can now subscribe or ignore boards on the fly. The changes are
not (yet) permanent. Writing the config file from the proxy is a thing I have to
think through twice, because it involves possible security risks. For now you still
have to change entropy.conf to make permanent changes to news_boards= or
news_ignore=

- Searching for news boards and message slots now uses HTLs of 2,4,8,16 for the
retries. This should - hopefully - reduce the number of message collisions.

- The german translations were updated after ages (de/entropy.lo)

- The FCP (src/client.c) now has the Pending/[URI=xxx]/EndMessage replies enabled
again. While this may break some FCP clients, it's still the clients that are
broken. I got both, FIW as well as the CVS version of goFrost working with the
reply. The big advandtage of the Pending replies is that they make it possible for
the node to detect closed client sockets much earlier. Before the file_get() would
loop until it was through all fragments, even if the client application's socket was
already closed.

- The FCP now sends HighestSeenBuild= with the NodeInfo reply, just as Freenet does.
FIW detects this and displays a warning if your build is outdated.

(thanks to pullmoll for doing change summaries now)