0.8.0-404.
Build 403 changes:
This one cleans up the FCP client.c code. It now uses the (long available)
sock_agets() function, which dynamically allocates buffer as a line is read in.
It also has the Pending messages sending an URI= line whenever an URI is known.

Other changes are a cleanup of the FLOCK semaphore code, introduced (page size) 4K
file regions to lock (for a try) and got rid of useless close/open/dup after the
first fork of fork2().

Build 404 changes:
The most important change is a bugfix for the proxy. There was a memory leak in
caching meta data, because a locally declared variabled shadowed the function scope
variable of the same name (which would have been used to free shared memory).