0.8.2-411.  Changes:
Build 411 is out - it fixes a bug in the determination of the best/poor
chunks of a fragment. The code was in fact zapping the best matching
chunks and advertizing the worst matching ones. This pushed up the data
transfer a lot, because a node would always delete what it actually
should hold for others to request.