Update to inkscape 0.39. Numerous changes. Too many to list here. See the ChangeLog.
Use xdg-dirs.
GCC 3 is now required. And g++ is used. (Because of this,
this won't work with NetBSD 1.6.2 unless dependencies
are built with gcc 3 too.)
Depend on gtkmm and popt. And also glib2 and pango (which
probably don't matter because gtk2 depends on them.)

Add patch, because some systems do not know fpresetsticky().
Hopefully, fpsetsticky is good enough.