Even if "Overwriting" record the data in the package database (db file).
Fix problem with the check for the file belonging to older (previous)
package. And also change verbose message about this.
(Note: it doesn't delete the file yet for some more testing.)

Don't use db's R_NOOVERWRITE flag when storing a value. Just default
to overwrite the data. This was required so we could easily see
if files from old/previous package should be removed or not.

Bump PKGREVISION.