Fix download site location.
Bump to 0.8.3 for misc fixes. Two original patches don't seem to be needed
anymore as the DNOTIFY code is now #ifdef'd away and the wireless applet
is conditionally built. zh_TW translation has crept in which breaks with
the system base gettext tools, so the pkgsrc ones override it. Two patches
are added to make sure those gettext tools are used rather than whatever is
found first in $PATH.