Initial import of rte-0.5.6:
The RTE library is a frontend or wrapper of other libraries or
programs for real time video and audio compression on Linux.
Currently it works on x86 CPUs only, sorry. The library is designed
to interface between codecs and the Zapping TV viewer:
http://zapping.sourceforge.net, precisely its recording plugin.

TODO:

Installs and deinstalls fine. Completely untested,
since zapping only uses it under Linux.