initial import of sflphoned
Voice over IP tends to be part of more and more corporate (and home) networks
today. As Open Source Software become more and more popular in companies, we
decided to provide the community with a new powerful and open platform for
desktop-based Voice over IP. SFLphone aims to become your desktop's VoIP
companion.