Another graphical front-end to CUPS, functionally very comparable to xpp. Both satisfy the goal of being command-line compatible with lp and lpr while presenting the user with *all* the printer-specific options documented in the PPD file, and using the UIConstraints in the PPD to notify the user if any conflicting selections have been made (xpp does this instantly, in a very slick and nondistracting way; GtkLP can be configured to check constraints always, or provide a button to check them and report any conflicts).
The kprinter tool of KDE also offers comparable functionality, provided
PKG_OPTIONS.kdelibs3=cups was present at build (and presumably also
PKG_OPTIONS.qt3-libs=cups and PKG_OPTIONS.qt3-tools=cups when those were
built as well). It notifies the user of conflicting selections with a dialog
at save/OK time; the user, on clicking Driver Settings (not Advanced as the
dialog instructs), can see which settings are in conflict. The xfprint4 tool
in XFCE seems to have missed the boat on providing access to CUPS features
at all. I haven't tried gnome, but by report it also misses the boat at
present, but may perhaps improve; meanwhile GtkLP or xpp can perhaps be used
to print from it.
http://mail.gnome.org/archives/usability/2005-December/msg00021.html

GtkLP has a companion GtkLPQ for showing queue status; xpp doesn't (but
with CUPS you can just browse http://localhost:631/ for that anyway).
GtkLP has better support for submitting a multiple-file job, better
handling of file names with whitespace or odd characters, and suppresses
the file selection when run as a filter (I wish xpp did this). xpp's
file selection offers a preview for several file types.

xpp is more faithful to the grouping of options in the PPD, presenting
them on separate tabs under the PPD's suggested readable names, and
correctly treats the InstallableOptions group as view-only; GtkLP tucks
them all away on an obscure "PPD" tab, and allows the user to frob the
InstallableOptions to values that don't match the printer's actual
capabilities, which could lead to confusion.

I was in a hurry to test this so I temporarily set IGNORE_RECOMMENDED to
avoid updating a bunch of gtk2 stuff. It works for me but someone might
check that nothing breaks with the more current gtk2.