Update pkgsrc/audio/audacity to 1.3.0 (beta version), fix some pkglint warnings, reindent, and import this as pkgsrc-wip/audacity-devel.
Audacity is a program that manipulates digital audio waveforms. It supports all
common editing operations such as Cut, Copy, and Paste, plus it will mix tracks
and apply plug-in effects to any part of a sound. It also has a built-in
amplitude envelope editor, a customizable spectrogram mode and a frequency
analysis window for audio analysis applications.