Initial commit.  Note: requires a not-yet-committed buildlink3.mk for pidgin (which has been sent to Jan Danielsson).
Guifications is a Pidgin plugin that displays "toaster" popups in a
user-defined corner of the screen, similar to features that have been
added to the official MSN Messenger (now called Windows Live Messenger),
Yahoo! Messenger and AOL Instant Messenger clients. It's highly
configurable, easy to use, and has theme support. It really is the end
all, be all, toaster popup plugin for Pidgin!