Initial import of silvertree-0.1 (an early alpha version):
A 3D turn-based RPG engine and adventure

The SilverTree Project will produce a 3D RPG with a hybrid real
time, turn based system.

SilverTree intends to include a wide range of adventuring activities,
in addition to combat and looting. In the wilderness tracking and
stealth can provide the all-important element of surprise in combat,
or allow the player to avoid more powerful foes. Literacy allows
characters to use books and manuscripts to study new skills.

Grinding will be minimized. Less experience is gained each time
the same type of enemy is defeated, and healing is not easily
obtained. Most enemies will try to flee from the PC when they have
little chance of survival.

SilverTree will allow play as several different types of heroes.
The PC's response to moral dilemmas will be measured on several
scales such as Honest <-> Deceptive, and NPCs will react accordingly,
but there is not only one right answer.