Update to 3.1:
Jun 16, 2007   Version 3.1 released. This is a bugfix release, no
features added. Fixed locale and .desktop installation directory
for Gentoo. Increased the speed of search engine indexing. Fixed
the problem when kchmviewer is called from a browser with URL like
file://. Fixed the unability to run the program when the last opened
file does not exist anymore, and the LoadLatestFileOnStartup is
set to 1. Fixed CHMs with Unicode index/topic filenames inside CHM
file. Fixed showing book images in QTextBrowser when file path
contains urlencoded symbols. Fixed 64bit platforms compilation if
system chmlib is not present.

The version 3.1 will be the last version which supports Qt3. It
will be supported for some time with bugfixes only. The next official
version will be 4.0, which will be based on Qt4.

Mar 14, 2007   Version 3.0 released. The most important improverment
was a new search engine, which could correctly search in any language
(including Asian languages), and can search for symbols - so feel
free searching for "new SWFActon()" in PHP help. Added DCOP bindings
for KDE version. Chm procesing code moved in a separate library
and was rewritten resulted in up to 20x(!) speed increase for
loading large chm files. Fixed a lot of old bugs, including 'Copy'
in KDE, removing a duplicated application title, and incorrect
popup menu position.