Initial import of opentracker CVS version while I wait for the official release:
This is opentracker. An open bittorrent tracker.

This tracker is open in a sense that everyone announcing a torrent
is welcome to do so and will be informed about anyone else announcing
the same torrent. Only source IPs are accepted. The tracker implements
a minimal set of essential features only but was able respond to
far more than 10000 requests per second on a Sun Fire 2200 M2.