Update to 1.4.6 (old master site and distfile are missing).  Changes:
1.4.0 introduces the new feature of usermaps, which are a way to let
users have temporary mail addresses which they can use when they deal
with possible spammers.

1.4.1 fixes some of the bugs that were in 1.4.0, adds missing features
(and the missing manual page for usermap(7).

1.4.3 introduces one new feature; if you set the config file option
junkfolder=, unwanted mail will not be bounced but will instead be put
into a junkmail folder.

1.4.4 adds spam=-like configuration options to manage connections from
blacklisted sites similar to how spam-ridden messages are currently
handled.

1.4.5 fixes a minor bug and has a fairly extensive code scrub to sweep
out things that the FSF's so-called 'C' compiler complains about.

1.4.6 adds a new configuration option - trusted=hostname, which tells
postoffice to treat hostname as if it was localhost.