Import fasm-1.67.27 as wip/fasm.
The flat assembler is a fast and efficient self-assembling 80x86 assembler
for DOS, Windows and Linux operating systems.

Currently it supports all 8086-80486/Pentium instructions with MMX, SSE,
SSE2, SSE3 and 3DNow! extensions and x86-64 (both AMD64 and EM64T)
instructions, can produce output in binary, MZ, PE, COFF or ELF format.
It includes the powerful but easy to use macroinstruction support and
does multiple passes to optimize the instruction codes for size.
The flat assembler is self-compilable and the full source code is included.