Import bsdbuild-102608 as wip/bsdbuild.
  __   __  _
 |__) |_  |  \
 |__) __| |_ /
   |   |   |
   +---+---+------> Build

This is BSDBuild, a self-contained, portable build system in the spirit
of the traditional 4.4BSD make library files. It allows standard BSD-style
makefiles to be used, but without the BSD make extensions (Bourne script
fragments are used instead), so it is portable to most operating systems
and make flavors.

In addition, BSDBuild includes a configure script compiler and a number
of software test modules.

BSDBuild is freely available from: http://hypertriton.com/bsdbuild/.