Import cpm-0.22beta (as cpm-0.22rc1):
This program is a ncurses based console tool to manage passwords and
store them public key encrypted in a file - even for more than one
person. The encryption is handled via GnuPG so the programs data can be
accessed via gpg as well, in case you want to have a look inside. The
data is stored as as zlib compressed XML so it's even possible to reuse
the data for some other purpose.