Import texlive-texmf-base-2008 as wip/texlive-texmf-base.
TeX Live is an easy way to get up and running with TeX.  It provides a
comprehensive TeX system with binaries for most flavors of Unix,
including GNU/Linux, and also Windows.  It includes all the major
TeX-related programs, macro packages, and fonts that are free
software, including support for many languages around the world.

This package contains essential texmf files of TeX Live.