Import icecat-3.0.4 as wip/icecat, largely based on www/firefox3.
Gnuzilla is the GNU version of the Mozilla suite, and GNU IceCat is the GNU
version of the Firefox browser. Its main advantage is an ethical one: it is
entirely free software. While the source code from the Mozilla project is free
software, the binaries that they release include additional non-free software.
Also, they distribute and recommend non-free software as plug-ins.

In addition, GNU IceCat includes some privacy protection features:

   1. Some sites refer to zero-size images on other hosts to keep track of
      cookies. When IceCat detects this mechanism it blocks cookies from the
      site hosting the zero-length image file. (It is possible to re-enable
      such a site by removing it from the blocked hosts list.)
   2. Other sites rewrite the host name in links redirecting the user to
      another site, mainly to "spy" on clicks. When this behavior is detected,
      IceCat shows a message alerting the user.