Import dragbox-0.4.0 as wip/dragbox.
Dragbox is a tool for connecting the command line with the desktop
environment. It summons a drag handle in a window when you are managing
files or text in the shell, connecting the different workspaces -- desktop
and command line. Dragbox can take items on the command line or from
a pipe and put on its shelf. The inverse is also possible:
dragged-to items can be output ,to the shell. It is possible
to have more than one instance running, and put items on or get items
from any of them. Right-click in the window to open a context menu
or access the preferences window. Clicking an item copies it to the clipboard
(files can be opened as an option).
For detailed command line instructions, please read the included man page.