Import newsbeuter-1.3 as wip/newsbeuter.
Newsbeuter is an open-source RSS/Atom feed reader for text terminals.
It runs on Linux, FreeBSD, Mac OS X and other Unix-like operating systems.
Newsbeuter's great configurability and vast number of features make it a
perfect choice for people that need a slick and fast feed reader that can
be completely controlled via keyboard. It's the Mutt of RSS feed readers.