Import tpp-1.3.1 as wip/tpp.
tpp stands for text presentation program and is an ncurses-based
presentation tool. The presentation can be written with your favorite
editor in a simple description format and then shown on any text
terminal that is supported by ncurses - ranging from an old VT100 to
the Linux framebuffer to an xterm.